Channel for each Directory
Changes how clients keep track of the FileSystem state. Each client now
connects to a channel specific to the directory they are in. Events
about file additions/deletions are broadcast to these channels. This
means clients don't have to keep track of all of the presence state -
and permissions can easily be added in future.