authentication framework
adds initial framework for authentication.

username/password authentication is used initially. The
username/passsword will be sent to the authentication servers - for now
we use fake data.

Once the user has authenticated with the authentication service, a JSON
Web Token is generated. This JWT can be verified by any of the file
servers. This JWT is used by the client when making API requests