improves conflict reslolution
When conflicting replicas existed, the conflict was resolved by choosing
the replica with the max hash - the hash is of the file's contents.

This changes the conflict resolution to instead group the replicas by
hash, and then choose a replica from the largest group.

It also adds a repair mechanism to remove the conflicting replicas. Each
read has a 10% chance to trigger the repair mechanism.